requirements.txt
requests~=2.32.3
PyYAML~=6.0.2
langchain-community~=0.3.15
langchain~=0.3.15
PyPDF2~=3.0.1
numpy~=2.2.1
sentence-transformers~=3.3.1
pandas~=2.2.3
reportlab~=4.2.5